backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.19
python-multipart==0.0.9
aiofiles==23.2.1
Pillow>=12.0.0
fal-client==0.5.6
python-dotenv==1.0.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
rembg[cpu]>=2.0.50
edge-tts>=7.0.0
google-generativeai>=0.8.0
pdfplumber>=0.11.0
